Discord bot/requirements.txt
hikari==2.1.1
hikari-lightbulb==2.3.5.post1
psutil==7.1.3
aiohttp==3.13.2
certifi==2025.11.12
